+++ lib/karafka/web/tracking/consumers/contracts/topic.rb
@@ -6,7 +6,7 @@ module Karafka
       module Consumers
         module Contracts
           # Expected topic information that needs to go out
-          class Topic < Tracking::Contracts::Base
+          class Topic < Web::Contracts::Base
             configure
 
             required(:name) { |val| val.is_a?(String) && !val.empty? }
@@ -17,7 +17,7 @@ module Karafka
 
               partition_contract = Partition.new
 
-              data.fetch(:partitions).each do |_partition_id, details|
+              data.fetch(:partitions).each_value do |details|
                 partition_contract.validate!(details)
               end