+++ lib/karafka/web/tracking/contracts/error.rb
@@ -7,7 +7,7 @@ module Karafka
         # Contract for error reporting
         # Since producers and consumers report their errors to the same topic, we need to have
         # a unified contract for both
-        class Error < Base
+        class Error < Web::Contracts::Base
           configure
 
           required(:schema_version) { |val| val.is_a?(String) }